server/requirements.txt
# MQTT
paho-mqtt>=1.6.1

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
python-multipart

# Database
sqlmodel==0.0.16
aiosqlite

# Authentication
python-jose[cryptography]
passlib[argon2]
firebase-admin

# Configuration
pydantic-settings
email-validator
python-dotenv

# AI/ML (STABLE SETUP)
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.8.0
joblib==1.4.2
tensorflow-model-optimization

# Data Processing (FORCE modern versions)
numpy==1.26.4
pandas==2.2.2
Pillow>=10.0.0

# AI Chatbot
google-generativeai
groq

# HTTP Client
httpx

# Payments
razorpay

# Rate limiting
slowapi
